Fix Tailwind build styling
I realized that addons can only use Tailwind styles that are used in the main app. This is due to the fact that Tailwind scans the app on build, and includes only those classes that it finds. I'm updating some of my styling here to rely on existing classes.

May be worth looking into a more dynamic solution if the addon ecosystem is going to grow.